Adding Spinners
Adding loading spinners throughout where they are needed... except for the menu. Updated state to include a  boolean check on whether images or videos exist for the active date, so we can load a spinner which the actual fetching is happening. Don't need it on videos, but it's in place in-case I get into the API.